drivers: touch and leds
gen_keys_720P and gen_keys_ist differ only in the functions for the device tree, so we will leave the native gen_keys_720. Remove the duplicate config TOUCH SCREEN_IST30XX. Also in our led show 2013 driver there is no tp_color, so we will remove it and fix the compilation.